Manager: Serialize file paths instead of file infos
Deserializing FileInfos with Newtonsoft.Json on Mono lead to invalid FileInfo objects (correct paths, wrong state).

It is also not supported on .net core, ISerializable implementations in FileInfo throw PlatformNotSupported exceptions and Newtonsoft.Json works around that. We may want to switch to a different serialization library in the future, so we want to avoid relying on that.