Analytics: Fix Deimos success detection
Finding the gadget which makes up the enemy below 10% of its health has been simplified and we now check whether its attack target ever becomes targetable. That always happen when players are teleported to its platform near the end of the encounter.

This was not checked before and sometimes the 10% gadget would go untargetable which resulted in false successes.