Analytics: Experimental Twisted Castle success detection
Not sure if checking for rewards works here if doing this twice in the same week, don't have enough logs to confirm that.

Reward id is 496, type 60685.

Some other potential alternatives if it doesn't work:
- Gliding Disabled (34298) is removed at once for all players on success, along with Madness (35006) stacks and Eligibility: Mildly Insane Achievement (34964) for living players (those are lost on death).

However, Gliding Disabled also seems to be removed on wipe, so checking if players are alive might be necessary (and prone to timing issues).