Analytics: Introduce health calculations and fix Deimos health
This moves health % calculations from the log manager into Analytics. This allows us use EncounterData to define special handling of bosses where the typical method (maximum of lowest amounts of health reached) does not work. An example of this would be Deimos, this fixes the health being unreliable in the Log Manager.