Analytics: Add success detection for kitty golems
Sigh. Not particularly happy that we have to go this way,
but there just doesn't seem any reliable way to detect
the results, so we use a health-based approach, similar
to what Elite Insights does.

Kitty golems logs are so important that we just cannot
afford to have most of them marked as "Unknown".