Analytics: Add API to do log parsing and processing in one step
This greatly reduces allocations, leading to a ~25% speedup when processing a batch of logs. When using the new overload in LogProcessor, we no longer construct a ParsedLog, and instead just use each parsed struct once.

There are also small improvements for doing both steps individually (reduced copying of big structs),
and the EVTC Inspector now supports both of these modes.

|                             Method |    Mean |    Error |   StdDev |
|----------------------------------- |--------:|---------:|---------:|
|   DirectProcess_NParser_NProcessor | 7.264 s | 0.0442 s | 0.0413 s |
| ParseAndProcess_NParser_NProcessor | 8.449 s | 0.0619 s | 0.0579 s |
| ParseAndProcess_NParser_OProcessor | 8.526 s | 0.0629 s | 0.0588 s |
| ParseAndProcess_OParser_OProcessor | 8.883 s | 0.0750 s | 0.0702 s |